Some improvements to CanWaitForCounter
This test didn't make much sence because there could be a race condition
right off the bat by calling counter async twice. The Canwaitforcounter
method looks at what ever the starting value is and waits, so if it runs
before the selected time, it will be waiting for nothing and fail.