Looks like we have to manage DLQ's for ASQ.
I've read a few blog posts and it sounds like the retry policy is more
connection based and doesn't dead letter or poison messages

https://alexandrebrisebois.wordpress.com/2013/08/14/poison-queues-are-a-must/

“To find "poison" messages in Storage queues, when dequeuing a message
the application examines the DequeueCount property of the message. If
DequeueCount is above a given threshold, the application moves the
message to an application-defined "dead letter" queue.”

"If the message processing exception was related to a service call, or
database operation, and there was a connection failure, the receiving
application will process the message again and may well succeed on the
second try. If, however, the exception was caused by the content of the
message, the application will receive the message again, attempt to
process it, catch an exception and abandon it, receive the message
again, attempt to process it, catch an exception and abandon it, receive
the message again, attempt to process it, catch an exception and abandon
it. If there is no mechanism to detect that this is a “poison message”
the cycle will continue indefinitely.