Upgrades to xUnit v3
Upgrades the test suite to use xUnit v3.

This change modernizes the testing framework and allows us to take advantage of the latest features and improvements in xUnit.
It also updates the test harness to remove the xunit dependency and the Foundatio.Xunit project to include xunit v3 dependencies and a description to clarify the usage with xUnit v2 via the Foundatio.Xunit2 project.