Upgrade to xunit v3, Adds Foundatio.Xunit.v3 (#431)
* Added Foundatio.Xunit2 package

* Upgrades to xUnit v3

Upgrades the test suite to use xUnit v3.

This change modernizes the testing framework and allows us to take advantage of the latest features and improvements in xUnit.
It also updates the test harness to remove the xunit dependency and the Foundatio.Xunit project to include xunit v3 dependencies and a description to clarify the usage with xUnit v2 via the Foundatio.Xunit2 project.

* Adds retry functionality for xUnit tests

Introduces attributes and associated classes to enable automatic retries for failed xUnit tests. This enhances test reliability by allowing tests to pass after transient failures.

The changes include:
- Introducing `RetryFactAttribute` and `RetryTheoryAttribute` to mark tests for retry.
- Implementing `RetryFactDiscoverer` and `RetryTheoryDiscoverer` to discover retry test cases.
- Creating `RetryTestCase` and `RetryTheoryTestCase` to handle the retry logic.
- Adding a `DelayedMessageBus` to buffer test messages until the final result is determined.
- Updating test logger classes to ValueTask.
- Adding an extension to get the current cancellation token.

These changes provide a mechanism to automatically retry flaky tests, reducing the need for manual intervention and improving overall test stability.

* Removes Xunit.Abstractions dependency

Removes the unnecessary dependency on Xunit.Abstractions in test harness and test projects. This simplifies the project structure and reduces potential conflicts without impacting functionality.

* Adds cancellation token support to tests

Updates tests to properly pass the cancellation token when calling async methods.
This ensures that the tests are cancellable and prevents potential deadlocks or timeouts.

* PR feedback

* PR Feedback

* Renames CancellationToken for test context

Renames the static CancellationToken property in TestWithLoggingBase to TestCancellationToken to provide a more descriptive and specific name for the test context cancellation token.
This change improves code clarity and avoids potential naming conflicts in test classes.
Also, it updates all usages of the property in test classes to use the new name.

* renamed xunit to xunit.v3

* Renames Foundatio.Xunit2 to Foundatio.Xunit

Renames the Foundatio.Xunit2 project to Foundatio.Xunit to align with xUnit v2.
Updates project references and namespaces accordingly.

This change prepares for removing xUnit v2 support and
consolidating the xUnit helpers into a single project.